fix(cli/clilog): address pipe error review
